requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
Pillow==11.3.0
python-decouple==3.8
psycopg2-binary==2.9.10
redis==6.2.0
celery==5.5.3
django-celery-beat==2.8.1
django-filter==25.1
django-extensions==3.2.3
django-storages==1.14.2
boto3==1.34.69
stripe==8.8.0
twilio==9.1.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0 